Fixed missing bone in rotation list. Fixed LookPose animation file - which seemed to have the wrong number of frames for an unknown reason. Cleaned up coding style for several of the classes which had come from GameSDK. Added in some debug code so I can follow what's happening and how. Cleaned up some of the fragment and scope settings in the controller def.
Noticed that the compromise made to get the character slot causes a failure in look ik when using the character editor and mannequin - as neither can access an CActorComponent in the entity. This was pretty much to be expected, but I need to find a way to get the character from the look procedural context. Using the built in animated character makes this a little trickier - since we need the slotID to be a set value for it to work with existing code.